# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/me:
    get:
      operationId: getMe
      tags: [users]
      summary: Get current user profile
      security:
        - bearerAuth: []
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
    patch:
      operationId: updateMe
      tags: [users]
      summary: Update current user profile
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserBody"
      responses:
        "200":
          description: Updated user profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
  /resumes:
    get:
      operationId: listResumes
      tags: [resumes]
      summary: List all resumes for current user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of resumes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Resume"
    post:
      operationId: createResume
      tags: [resumes]
      summary: Create a new resume
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateResumeBody"
      responses:
        "201":
          description: Created resume
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Resume"
  /resumes/{id}:
    get:
      operationId: getResume
      tags: [resumes]
      summary: Get a resume by ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Resume
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Resume"
    put:
      operationId: updateResume
      tags: [resumes]
      summary: Update a resume
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateResumeBody"
      responses:
        "200":
          description: Updated resume
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Resume"
    delete:
      operationId: deleteResume
      tags: [resumes]
      summary: Delete a resume
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
...
    UpdateUserBody:
      type: object
      properties:
        name:
          type: string
        avatarUrl:
          type: string
    UserProfile:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
        name:
          type: string
        avatarUrl:
          type: string
        plan:
          type: string
          enum: [free, pro]
        usageResumeCount:
          type: integer
        usageCoverLetterCount:
          type: integer
        subscriptionExpiresAt:
          type: string
          nullable: true
        createdAt:
          type: string
      required:
        - id
        - email
        - plan
        - usageResumeCount
        - usageCoverLetterCount
    WorkExperience:
      type: object
      properties:
        jobTitle:
          type: string
        company:
          type: string
        location:
          type: string
        startDate:
          type: string
        endDate:
          type: string
        isCurrent:
          type: boolean
        responsibilities:
          type: string
        bullets:
          type: array
          items:
            type: string
      required:
        - jobTitle
        - company
        - startDate
    Education:
      type: object
      properties:
        degree:
          type: string
        institution:
          type: string
        location:
          type: string
        graduationYear:
          type: string
        gpa:
          type: string
      required:
        - degree
        - institution
    ResumeData:
      type: object
      properties:
        personalInfo:
          type: object
          properties:
            fullName:
              type: string
            email:
              type: string
            phone:
              type: string
            location:
              type: string
            website:
              type: string
            linkedin:
              type: string
          required:
            - fullName
            - email
        summary:
          type: string
        workExperience:
          type: array
          items:
            $ref: "#/components/schemas/WorkExperience"
        education:
          type: array
          items:
            $ref: "#/components/schemas/Education"
        skills:
          type: array
          items:
            type: string
        template:
          type: string
          enum: [modern, classic, minimal, creative]
      required:
        - personalInfo
    Resume:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        title:
          type: string
        data:
          $ref: "#/components/schemas/ResumeData"
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - userId
        - title
        - data
        - createdAt
        - updatedAt
    CreateResumeBody:
      type: object
      properties:
        title:
          type: string
        data:
          $ref: "#/components/schemas/ResumeData"
      required:
        - title
        - data
...